+++ src/components/index.ts
@@ -1,4 +1,4 @@
 export { default as CollectionScopesMatrix } from './CollectionScopesMatrix'
 export { default as GlobalScopesMatrix } from './GlobalScopesMatrix'
-export { OAuthConnectBanner } from './OAuthConnectBanner'
+export { AgentConnectPill } from './AgentConnectPill'
 export { OAuthView } from './OAuthView'